Revert setting default value for DocType Element
This reverts 3304085e3650e339b961404624983af772ea31d5.

The default value was originally unset. Commit 3304085 changed it to
"matroska" reasoning that the Matroska specs (which listed "matroska" as
the default value) were maintained better.

However: as pointed out in #74 the Matroska specs state why they have
the default differ from the EBML spec. Quoting the Matroska specs:

> The default values defined for the EBML header correspond to the
> values for a Matroska stream/file.  When parsing the EBML header the
> default values are different, irrespective of the DocType defined.

Therefore it stands to reason that the EBML spec should really stay with
not having a default value for DocType.

Fixes #74.