[RFC] Explicit the EBML parent of an element
With the level-based definition of an element we rely on the order of which the elements are defined. It also makes it tricky to define elements in a separate specification (think WebM elements, DivX elements, or deprecated elements that won't make it to the RFC).

I think we should set the explicit parent for each element. That means also including the parents of the parent too.
For example ChapterTimeStart: Chapters.EditionEntry.ChapterAtom