Relax EBML Header length restrictions
libavformat (ffmpeg's multiplexer) has always coded the size field of the EBML element on eight bytes for all Matroska and Webm files it was used to mux. The current requirement of coding every size in the header on four bytes or less would therefore invalidate lots of files which makes it unacceptable.

As this requirement was stated twice, I have removed one occurence of it. The description of EBML Body doesn't repeat the restrictions imposed by EBMLMaxSize/IDLength either.