split the definition of the elements ending a unknown size element
The "not a valid Descendant" was allowing non valid elements to be fine with
ending an unknown size elements. But it shouldn't.

A parent element is easy to figure out on its own and is likely the most common case.
The other case can only be a higher level element.

And it's not the beginning of the element that ends the Unknown-Size element
(what beginning ? just the ID, the ID+Size ?).