make the end of a file an end of EBML Document
Concatenated EBML Document also ends the current unknown size element, even though
the file is not finished. A EBML Header could be considered a non Descendant Element
but it's not really part of the Schema so it's better to make it clearer and having
to dig into that gray area.